feat: create driven adapter to use database

+++ domain/model/src/main/java/co/com/pragma/model/loan_type/gateways/LoanTypeRepository.java
@@ -3,5 +3,5 @@ package co.com.pragma.model.loan_type.gateways;
 import reactor.core.publisher.Mono;
 
 public interface LoanTypeRepository {
-    Mono<Boolean> existsById(Integer id);
+    Mono<Boolean> existsById(Long id);
 }
