+++ domain/model/src/main/java/co/com/pragma/model/state/gateways/StateRepository.java
@@ -3,5 +3,5 @@ package co.com.pragma.model.state.gateways;
 import reactor.core.publisher.Mono;
 
 public interface StateRepository {
-    Mono<Boolean> existsById(Integer id);
+    Mono<Boolean> existsById(Long id);
 }
